+++ Pholors/EndGameVC.h
@@ -6,12 +6,11 @@
 //  Copyright (c) 2013 Rock Bottom. All rights reserved.
 //
 
-
 //Responsable for the view that shows up when the time ends in T.A.
 
 @interface EndGameVC : UIViewController
 
-@property int points;
+@property NSInteger points;
 @property BOOL highscore;
 
 @end
